Fix code review issues from text block deep linking PR #958
- Gate cross-document source clicks to navigate-only, avoiding flash of
  local selection state before navigation (CorpusChat.tsx)
- Extract useClearTextBlockOnInteraction hook to deduplicate clearing
  logic and hoist it from per-page PDFPage to parent PDF component,
  eliminating O(visible-pages) redundant navigate calls
- Fix buildQueryParams to delete ?tb= when textBlock is null
- Add LocationLike/NavigateFn type aliases for all update*Param functions
